Create Scene Loader
Loads a scene with the name entered in the script's serialize field. Kinda annoying that it uses on trigger enter, as ideally I'd spawn the layer on the trigger when entering a new scene but then it just instantly teleports them to the previous scene. Plus it sometimes causes duplicate player characters to exist at the same time, which is strange