Create Scene Transfer Player Positioner
Since the same player game object is transferred between every scene, this code teleports them to the 'Start' game object upon a scene transition. Bonus of using the same player game object is that the players velocity is transferred. Currently a WIP since it also tps the player upon starting the game, which it isn't supposed to do.